requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
boto3==1.24.61
botocore==1.27.61
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==2.2.5
Django==4.1
django-ckeditor==6.5.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-editorjs==0.2.1
django-extensions==3.2.0
django-ipware==4.0.2
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-quill-editor==0.1.40
django-rest-framework==0.1.0
django-solo==2.0.0
django-storages==1.13.1
django-tables2==2.4.1
django-timezone-field==5.0
djangorestframework==3.13.1
drf-generators==0.5.0
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.9.1
drf-yasg==1.21.3
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.15.0
MarkupSafe==2.1.1
packaging==21.3
phonenumbers==8.12.53
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
pydantic==1.9.2
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
pyzoom==0.1.15
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
razorpay==1.3.0
